requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pydantic==1.10.13
PyYAML==6.0.3
typer==0.9.0
httpx==0.24.1
beautifulsoup4==4.12.2
rapidfuzz==3.5.2
